Backend/requirements.txt
asgiref==3.8.1
Authlib==1.3.1
azure-core==1.30.2
azure-storage-blob==12.19.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==4.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.14.0
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pycparser==2.22
python-dotenv==0.21.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
pillow==10.4.0
drf-yasg==1.21.7
